User dataclass created
I was intending to make this a proper bean, but upon googling it seems
there is some division on whether it is considered good style, hence I
pick the easier option for now and will make the actual choice later.

Adding a getKey method which for the moment returns the simplest key
which is the email. I do this as I forsee the need for avoiding
collisions.

Adding a isValid method to make sure there is some data in all of the
fields.