Using User, adding internal list of users and POST
The users are stored in an internal state as a HashMap as that makes
lookup easy. The users are assumed to uniquely identified by their
key so this makes avoiding duplicates simple.

POST requests will if they succeed result in a user being added to
users. It may fail for the following reasons:
- User is not valid
- User already exists
- User has a key that does not form a valid URI
This final requirement should be avoided by design of the key.